double to big (#20)

+++ src/main/java/com/transferwise/acorn/models/BalanceResponse.java
@@ -2,6 +2,7 @@ package com.transferwise.acorn.models;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -18,6 +19,6 @@ public class BalanceResponse {
     public List<Step> steps;
     public SourceAmount sourceAmount;
     public TargetAmount targetAmount;
-    public double rate;
+    public BigDecimal rate;
     public List<FeeAmount> feeAmounts;
 }
